move Hostname struct
